+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- Added development builds to npm registry (`npm i uikit@dev` / `yarn add uikit@dev`)
+- Add development builds to npm registry (`npm i uikit@dev` / `yarn add uikit@dev`)
 
 ### Fixed
 
@@ -14,6 +14,7 @@
 - Fix height of Masonry Grid
 - Fix attribute change listener
 - Cover component updates on image load
+- Fix Lightbox options handling 
 
 ## 3.0.0 rc 5 (June 4, 2018)